refactor(catalog): move the album components to a rightbar-agnostic home
The album card and its controls are shared surface for the modal route, not
rightbar furniture; they now live under catalog/album with their tests
mirrored alongside. The e2e page object targets the modal dialog instead of
the rightbar sidebar.
